Move raw Content::setFromPost() into ContentValuesTrait
